webhook-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
dhanhq==2.0.2
python-dotenv==1.0.1
pyotp==2.9.0
requests==2.32.3
playwright==1.48.0
aiohttp==3.10.11
